Display: Создан класс CommandFrame (насл Frame)
